fpgx35: work around GDI bug
GDI ExtFloodFill tries to convert a device specific bitmap to a normal bitmap for speed.

Prior to NT 4 this function has a bug, and so it will delete the device surface, which causes crashes and thus bugcheck (as it brings down csrss).

Do some pattern matching in DrvCopyBits and make it fail if it's being called by this conversion function. This causes a fallback to using the device specific bitmap, which is what is wanted.

Fixes bugcheck under NT 3.51 when running (at least) VC4 IDE.